+++ lib/_http_server.js
@@ -213,18 +213,11 @@ ServerResponse.prototype.writeHeader = function() {
 };
 
 
-var TLS_ARGS = ['pfx', 'key', 'passphrase', 'cert', 'ca', 'crl', 'ciphers',
-                'handshakeTimeout', 'honorCipherOrder', 'requestCert',
-                'rejectUnauthorized', 'NPNProtocols', 'SNICallback',
-                'sessionIdContext', 'secureProtocol', 'secureOptions'];
-
 function Server(options, requestListener) {
-  if (util.isObject(options)) {
-    // Only return an HTTPS server if valid TLS args are passed
-    for (var i = 0; i < TLS_ARGS.length; i++) {
-      if (options.hasOwnProperty(TLS_ARGS[i]))
-        return new require('https').Server(options, requestListener);
-    }
+  if (options !== null && typeof options === 'object') {
+    // Only return an HTTPS server if options.tls is true
+    if (options.tls)
+      return new require('https').Server(options, requestListener);
   } else {
     requestListener = options;
   }
