--- test/simple/test-https-from-http.js
@@ -1,69 +1,0 @@
-// Copyright Joyent, Inc. and other Node contributors.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a
-// copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to permit
-// persons to whom the Software is furnished to do so, subject to the
-// following conditions:
-//
-// The above copyright notice and this permission notice shall be included
-// in all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
-// NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
-// OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
-// USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-if (!process.versions.openssl) {
-  console.error('Skipping because node compiled without OpenSSL.');
-  process.exit(0);
-}
-
-var common = require('../common');
-var assert = require('assert');
-var fs = require('fs');
-var http = require('http');
-var https = require('https');
-
-var options = {
-  key: fs.readFileSync(common.fixturesDir + '/keys/agent1-key.pem'),
-  cert: fs.readFileSync(common.fixturesDir + '/keys/agent1-cert.pem')
-};
-
-var reqCount = 0;
-var body = 'test';
-
-var server = http.createServer(options, function(req, res) {
-  reqCount++;
-  res.writeHead(200, {'content-type': 'text/plain'});
-  res.end(body);
-});
-
-assert(server instanceof https.Server);
-
-server.listen(common.PORT, function() {
-  https.get({
-    port: common.PORT,
-    rejectUnauthorized: false
-  }, function(res) {
-    var data = '';
-
-    res.on('data', function(chunk) {
-      data += chunk.toString();
-    });
-
-    res.on('end', function() {
-      assert.equal(data, body);
-      server.close();
-    });
-  });
-});
-
-process.on('exit', function() {
-  assert.equal(1, reqCount);
-});
